added authorization and protect to tables routers

+++ app/tables/tables.controller.js
@@ -5,10 +5,10 @@ import asyncHandler from "express-async-handler";
 //@route POST/api/tables/
 //@access Public
 export const createTable = asyncHandler(async (req, res) => {
-    const {users_id, title, description} = req.body
+    const {title, description} = req.body
     const table = await prisma.tables.create({
         data: {
-            users_id, title, description
+            users_id: req.user.id, title, description
         }
     })
     res.json(table)
@@ -19,15 +19,17 @@ export const createTable = asyncHandler(async (req, res) => {
 //@access Public
 export const changeTable = asyncHandler(async (req, res) => {
     const {id, title, description} = req.body
-    const table = await prisma.tables.update({
-        where: {
-            id
-        },
-        data: {
-            title, description
-        }
-    })
-    res.json(table)
+    if(req.user.id){
+        const table = await prisma.tables.update({
+            where: {
+                id
+            },
+            data: {
+                title, description
+            }
+        })
+        res.json(table)
+    }
 })
 
 //@desc delete table
@@ -35,22 +37,24 @@ export const changeTable = asyncHandler(async (req, res) => {
 //@access Public
 export const deleteTable = asyncHandler(async (req, res) => {
     const {id} = req.body
-    const table = await prisma.tables.delete({
-        where: {
-            id
-        }
-    })
-    res.json(table)
+    if(req.user.id){
+        const table = await prisma.tables.delete({
+            where: {
+                id
+            }
+        })
+        res.json(table)
+    }
 })
 
 //@desc get tables
 //@route GET/api/tables/
 //@access Public
 export const getTables = asyncHandler(async (req, res) => {
-    const {users_id} = req.body
+    // const {users_id} = req.body
     const table = await prisma.tables.findMany({
         where: {
-            users_id
+            users_id : req.user.id
         }
     })
     res.json(table)
